feat: add option to open post on click if not in laptop
