+++ packages/shared/src/components/cards/WelcomePostCard.tsx
@@ -1,6 +1,6 @@
 import React, { forwardRef, ReactElement, Ref, useRef } from 'react';
 import classNames from 'classnames';
-import { CardButton, FreeformCardTitle, getPostClassNames } from './Card';
+import { FreeformCardTitle, getPostClassNames } from './Card';
 import ActionButtons from './ActionButtons';
 import { Container, generateTitleClamp, PostCardProps } from './common';
 import OptionsButton from '../buttons/OptionsButton';
@@ -10,9 +10,9 @@ import { Squad } from '../../graphql/sources';
 import { ActionType } from '../../graphql/actions';
 import FeedItemContainer from './FeedItemContainer';
 import { PostType } from '../../graphql/posts';
-import { useFeedPreviewMode } from '../../hooks';
 import { SquadPostCardHeader } from './common/SquadPostCardHeader';
 import { usePostImage } from '../../hooks/post/usePostImage';
+import CardOverlay from './common/CardOverlay';
 
 export const WelcomePostCard = forwardRef(function SharePostCard(
   {
@@ -34,7 +34,6 @@ export const WelcomePostCard = forwardRef(function SharePostCard(
   const { pinnedAt, type: postType } = post;
   const onPostCardClick = () => onPostClick(post);
   const containerRef = useRef<HTMLDivElement>();
-  const isFeedPreview = useFeedPreviewMode();
   const image = usePostImage(post);
   const { openStep, isChecklistVisible } = useSquadChecklist({
     squad: post.source as Squad,
@@ -61,9 +60,7 @@ export const WelcomePostCard = forwardRef(function SharePostCard(
       ref={ref}
       flagProps={{ pinnedAt }}
     >
-      {!isFeedPreview && (
-        <CardButton title={post.title} onClick={onPostCardClick} />
-      )}
+      <CardOverlay post={post} onPostCardClick={onPostCardClick} />
 
       <OptionsButton
         className="absolute right-2 top-2 group-hover:flex laptop:hidden"
